better Cache and KVStorage

+++ src/main/java/com/rebuild/server/service/base/GeneralEntityService.java
@@ -134,7 +134,7 @@ public class GeneralEntityService extends ObservableService  {
 		final ID currentUser = Application.getCurrentUser();
 
 		RecycleStore recycleBin = null;
-		if (SysConfiguration.getLong(ConfigurableItem.RecycleBinKeepingDays) > 0) {
+		if (SysConfiguration.getInt(ConfigurableItem.RecycleBinKeepingDays) > 0) {
 			recycleBin = new RecycleStore(currentUser);
 		}
